# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GrabGo API
  description: |
    Backend API for GrabGo food delivery platform.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## User Roles
    - **customer**: Regular users who order food
    - **restaurant**: Restaurant owners/managers
    - **rider**: Delivery riders
    - **admin**: Platform administrators
  version: 1.0.0
  contact:
    name: GrabGo Support
  license:
    name: ISC
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.grabgo.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/register endpoints
  schemas:
...
    Notification:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
        type:
          type: string
          enum: [order, promo, update, system, comment_reply, comment_reaction, referral_completed, payment_confirmed, delivery_arriving, milestone_bonus]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        data:
          type: object
        actors:
          type: array
          items:
            type: object
            properties:
              actorId:
                type: string
              actorName:
                type: string
              actorAvatar:
                type: string
        actorCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    NotificationSettings:
      type: object
      properties:
        chatMessages:
          type: boolean
          default: true
        orderUpdates:
          type: boolean
          default: true
        promoNotifications:
          type: boolean
          default: true
        commentReplies:
          type: boolean
          default: true
        commentReactions:
          type: boolean
          default: true
        referralUpdates:
          type: boolean
          default: true
        paymentUpdates:
          type: boolean
          default: true
        deliveryUpdates:
          type: boolean
          default: true
        systemUpdates:
          type: boolean
          default: true
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get user notifications
      description: Get paginated list of user's notifications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      hasMore:
                        type: boolean
                      totalPages:
                        type: integer
  /notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Get unread notification count
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unread count retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
        '404':
          description: Notification not found
...
  /users/settings/notifications:
    get:
      tags: [Notification Settings]
      summary: Get user's notification settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  notificationSettings:
                    $ref: '#/components/schemas/NotificationSettings'

    patch:
      tags: [Notification Settings]
      summary: Update user's notification settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              properties:
                settings:
                  type: object
                  properties:
                    chatMessages:
                      type: boolean
                    orderUpdates:
                      type: boolean
                    promoNotifications:
                      type: boolean
                    commentReplies:
                      type: boolean
                    commentReactions:
                      type: boolean
                    referralUpdates:
                      type: boolean
                    paymentUpdates:
                      type: boolean
                    deliveryUpdates:
                      type: boolean
                    systemUpdates:
                      type: boolean
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  notificationSettings:
                    $ref: '#/components/schemas/NotificationSettings'
        '400':
          description: Invalid settings format

  # ============ HEALTH ============
...